+++ cli/src/pcluster/resources/login_node/user_data.sh
@@ -35,6 +35,10 @@ export NO_PROXY="localhost,127.0.0.1,169.254.169.254"
 PROXY
 fi
 
+if [ "${DisableSudoAccessForDefault}" == "true" ]; then
+  sed -n -i "/"${OSUser}" ALL=(ALL) NOPASSWD:ALL/d" /etc/sudoers
+fi
+
 --==BOUNDARY==
 Content-Type: text/cloud-config; charset=us-ascii
 MIME-Version: 1.0
@@ -44,6 +48,9 @@ package_upgrade: false
 repo_upgrade: none
 
 datasource_list: [ Ec2, None ]
+
+${DisableSudoAccessForDefaultUserConfig}
+
 output:
   all: "| tee -a /var/log/cloud-init-output.log | logger -t user-data -s 2>/dev/console"
 write_files: